+++ src/routes/settings.js
@@ -1036,28 +1036,28 @@ const Settings = () => {
                 let badge = null;
                 let badgeName = null;
                 if (['lead_developer', 'project_manager', 'community_manager', 'development_team', 'support_manager', 'marketing_manager', 'support_team', 'marketing_team', 'graphic_team'].includes(sr.role)) {
-                    badge = <Tooltip key={`badge-${vars.userInfo.uid}-chub}`} placement="top" arrow title="CHub Staff"
+                    badge = <Tooltip key={`badge-${vars.userInfo.uid}-chub}`} placement="top" arrow title="CHub Team"
                         PopperProps={{ modifiers: [{ name: "offset", options: { offset: [0, -10] } }] }}>
                         <FontAwesomeIcon icon={faScrewdriverWrench} style={{ color: "#2fc1f7" }} />
                     </Tooltip>;
                     badgeName = "chub";
                 }
                 if (['community_legend'].includes(sr.role)) {
-                    badge = <Tooltip key={`badge-${vars.userInfo.uid}-legend`} placement="top" arrow title="CHub Community Legend"
+                    badge = <Tooltip key={`badge-${vars.userInfo.uid}-legend`} placement="top" arrow title="Community Legend"
                         PopperProps={{ modifiers: [{ name: "offset", options: { offset: [0, -10] } }] }}>
                         <FontAwesomeIcon icon={faCrown} style={{ color: "#b2db80" }} />
                     </Tooltip>;
                     badgeName = "legend";
                 }
                 if (['network_partner'].includes(sr.role)) {
-                    badge = <Tooltip key={`badge-${uid}-network-partner`} placement="top" arrow title="CHub Network Partner"
+                    badge = <Tooltip key={`badge-${uid}-network-partner`} placement="top" arrow title="Network Partner"
                         PopperProps={{ modifiers: [{ name: "offset", options: { offset: [0, -10] } }] }}>
                         <FontAwesomeIcon icon={faEarthAmericas} style={{ color: "#5ae9e1" }} />
                     </Tooltip>;
                     badgeName = "legend";
                 }
                 if (['platinum_sponsor', 'gold_sponsor', 'silver_sponsor', 'bronze_sponsor', 'server_booster', 'translation_team', 'web_client_beta'].includes(sr.role)) {
-                    badge = <Tooltip key={`badge-${vars.userInfo.uid}-supporter`} placement="top" arrow title="CHub Supporter"
+                    badge = <Tooltip key={`badge-${vars.userInfo.uid}-supporter`} placement="top" arrow title="Supporter"
                         PopperProps={{ modifiers: [{ name: "offset", options: { offset: [0, -10] } }] }}>
                         <FontAwesomeIcon icon={faClover} style={{ color: "#f47fff" }} />
                     </Tooltip>;
@@ -1396,7 +1396,7 @@ const Settings = () => {
                                         {vars.userInfo.name}
                                     </Typography>
                                     <Typography variant="h7" sx={{ flexGrow: 1, display: 'flex', alignItems: "center", maxWidth: "fit-content" }}>
-                                        {badges.map((badge, index) => { return <span key={index}>{badge}&nbsp;</span>; })}
+                                        {badges.map((badge, index) => { return <a key={index} onClick={() => { navigate("/badges"); }} style={{ cursor: "pointer" }}>{badge}&nbsp;</a>; })}
                                         {vars.userInfo.userid !== null && vars.userInfo.userid !== undefined && vars.userInfo.userid >= 0 && <Tooltip placement="top" arrow title="User ID"
                                             PopperProps={{ modifiers: [{ name: "offset", options: { offset: [0, -10] } }] }}><Typography variant="body2"><FontAwesomeIcon icon={faHashtag} />{vars.userInfo.userid}</Typography></Tooltip>}
                                     </Typography>